checking-code/pyproject.toml
[project]
name = "checking-code"
version = "0.1.0"
description = ""
authors = [
    {name = "HuleNet",email = "sidikovr04@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.121.1,<0.122.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "ruff (>=0.14.4,<0.15.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.17.1,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "bcrypt (==4.0.1)",
    "passlib (>=1.7.4,<2.0.0)",
    "itsdangerous (>=2.2.0,<3.0.0)"
]

[tool.poetry]
packages = [{include = "checking_code", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project.scripts]
app = "checking_code.main:start"
